requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
torch
transformers
altair
python-dotenv
requests
matplotlib
beautifulsoup4
lxml
huggingface_hub
hf_xet
tensorflow-cpu